pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbt-cost-guard"
version = "0.1.0"
description = "A CLI wrapper for dbt that estimates Snowflake query costs and requires confirmation for expensive runs"
authors = [
    {name = "dbt Cost Guard Contributors"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["dbt", "snowflake", "cost", "estimation", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "dbt-core>=1.5.0",
    "dbt-snowflake>=1.5.0",
    "click>=8.0.0",
    "snowflake-connector-python>=3.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.scripts]
dbt-cost-guard = "dbt_cost_guard.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/dbt-cost-guard"
Repository = "https://github.com/yourusername/dbt-cost-guard"

[tool.setuptools.packages.find]
where = ["."]
include = ["dbt_cost_guard*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=dbt_cost_guard",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "integration: marks tests as integration tests requiring Snowflake connection",
    "slow: marks tests as slow running",
]

[tool.coverage.run]
source = ["dbt_cost_guard"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

